Stock的最大交易数、最小交易数改为double类型

+++ hikyuu_cpp/hikyuu/Stock.h
@@ -110,13 +110,13 @@ public:
     int precision() const;
 
     /** 获取最小交易数量，同minTradeNumber */
-    size_t atom() const;
+    double atom() const;
 
     /** 获取最小交易数量 */
-    size_t minTradeNumber() const;
+    double minTradeNumber() const;
 
     /** 获取最大交易量 */
-    size_t maxTradeNumber() const;
+    double maxTradeNumber() const;
 
     /**
      * 获取指定时间段[start,end)内的权息信息
@@ -247,8 +247,8 @@ struct HKU_API Stock::Data {
     price_t m_tickValue;
     price_t m_unit;
     int m_precision;
-    size_t m_minTradeNumber;
-    size_t m_maxTradeNumber;
+    double m_minTradeNumber;
+    double m_maxTradeNumber;
 
     unordered_map<string, KRecordList*> pKData;
     unordered_map<string, std::shared_mutex*> pMutex;
@@ -256,7 +256,7 @@ struct HKU_API Stock::Data {
     Data();
     Data(const string& market, const string& code, const string& name, uint32_t type, bool valid,
          const Datetime& startDate, const Datetime& lastDate, price_t tick, price_t tickValue,
-         int precision, size_t minTradeNumber, size_t maxTradeNumber);
+         int precision, double minTradeNumber, double maxTradeNumber);
 
     virtual ~Data();
 };
